requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
azure-core==1.31.0
azure-functions==1.20.0
azure-servicebus==7.12.2
